ffssef

+++ fractol.h
@@ -6,7 +6,7 @@
 /*   By: gboudrie <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/25 17:27:05 by gboudrie          #+#    #+#             */
-/*   Updated: 2016/06/14 22:39:49 by gboudrie         ###   ########.fr       */
+/*   Updated: 2016/06/16 21:30:47 by gboudrie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@ typedef struct	s_env
 	int			bit;
 	int			end;
 	int			zoom;
-	int			x_decal;
-	int			y_decal;
+	double		x_decal;
+	double		y_decal;
 	int			param;
 	int			pos_x;
 	int			pos_y;
